Palm_detection_code/requirements.txt
calflops==0.3.2
grad_cam==1.5.2
matplotlib==3.6.2
mmcv==2.2.0
mmengine==0.10.6
numpy==1.24.1
opencv_python==4.9.0.80
pandas==2.2.3
Pillow==9.4.0
Pillow==11.1.0
pycocotools==2.0.7
pynvml==12.0.0
pyproj==3.7.0
PyYAML==6.0.1
PyYAML==6.0.2
rasterio==1.4.3
seaborn==0.13.2
Shapely==2.0.7
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.4
ultralytics==8.3.38
